refactor: reorganize realtime example and split audio judge helpers
- Move realtime directory from tests to examples/realtime-demo
- Split wrapJudgeForAudio into transcription and extraction variants
- Update imports and documentation references
- Clean up test helper organization
