+++ javascript/examples/realtime-demo/agents/vegetatrian-recipe.agent.ts
@@ -9,6 +9,7 @@ Your role is to:
 - Provide complete recipes with ingredients and step-by-step instructions
 - Keep responses concise and conversational for voice interaction
 - Be encouraging and enthusiastic about vegetarian cooking
+- Provide all of the ingredients and instructions for the recipe
 
 Remember:
 - This is a VOICE conversation, so speak naturally
