+++ javascript/examples/realtime-demo/realtime-client/components.json
