+++ javascript/examples/realtime-demo/realtime-client/src/components/ui/VoiceOrb.tsx
