+++ javascript/examples/realtime-demo/realtime-client/src/lib/utils.ts
