+++ javascript/examples/vitest/tests/helpers/index.ts
@@ -6,4 +6,5 @@ export {
   getAudioSegments,
 } from "./audio-conversation";
 export { OpenAiVoiceAgent } from "./openai-voice-agent";
-export { wrapJudgeForAudio } from "./wrap-judge-for-audio";
+export { wrapJudgeForAudioTranscription } from "./wrap-judge-for-audio-transcription";
+export { wrapJudgeForAudioExtraction } from "./wrap-judge-for-audio-extraction";
